requirements.txt
groq
pandas
python-dotenv
python-telegram-bot>=21.10,<22
